Move Glueby::Wallet into Glueby::Internal package
The wallet module should be used as an internal module.
