+++ lib/glueby.rb
@@ -2,5 +2,5 @@ require "glueby/version"
 
 module Glueby
   autoload :Contract, 'glueby/contract'
-  autoload :Wallet, 'glueby/wallet'
+  autoload :Internal, 'glueby/internal'
 end
