+++ lib/glueby/internal.rb
@@ -1,0 +1,5 @@
+module Glueby
+  module Internal
+    autoload :Wallet, 'glueby/internal/wallet'
+  end
+end
