+++ lib/glueby/internal/wallet/abstract_wallet_adapter.rb
@@ -1,0 +1,117 @@
+module Glueby
+  module Internal
+    class Wallet
+      # This is an abstract class for wallet adapters. If you want to use a wallet
+      # component with Glueby modules, you can use it by adding subclass of this abstract
+      # class for the wallet component.
+      class AbstractWalletAdapter
+        # Creates a new wallet inside the wallet component and returns `wallet_id`. The created
+        # wallet is loaded from at first.
+        # @return [String] wallet_id
+        def create_wallet
+          raise NotImplementedError, "You must implement #{self.class}##{__method__}"
+        end
+
+        # Delete the wallet inside the wallet component.
+        # This method expect that the wallet will be removed completely
+        # in the wallet and it cannot restore. It is assumed that the main use-case of
+        # this method is for testing.
+        #
+        # @param [String] wallet_id - The wallet id that is offered by `create_wallet()` method.
+        # @return [Boolean] The boolean value whether the deletion was success.
+        def delete_wallet(wallet_id)
+          raise NotImplementedError, "You must implement #{self.class}##{__method__}"
+        end
+
+        # Load the wallet inside the wallet component.
+        # This method makes the wallet to possible to use on other methods like `balance`,
+        # `list_unspent`, etc. All the methods that gets `wallet_id` as a argument needs to
+        # load the wallet before use it.
+        # If the wallet component doesn't need such as load operation to prepare to use the
+        # wallet, `load_wallet` can be empty method.
+        #
+        # @param [String] wallet_id - The wallet id that is offered by `create_wallet()` method.
+        # @return [Boolean] The boolean value whether the loading was success.
+        def load_wallet(wallet_id)
+          raise NotImplementedError, "You must implement #{self.class}##{__method__}"
+        end
+
+        # Unload the wallet inside the wallet component.
+        #
+        # @param [String] wallet_id - The wallet id that is offered by `create_wallet()` method.
+        # @return [Boolean] The boolean value whether the unloading was success.
+        def unload_wallet(wallet_id)
+          raise NotImplementedError, "You must implement #{self.class}##{__method__}"
+        end
+
+        # Returns list of loaded wallet_id.
+        #
+        # @return [Array of String] - Array of wallet_id
+        def wallets
+          raise NotImplementedError, "You must implement #{self.class}##{__method__}"
+        end
+
+        # Returns amount of tapyrus that the wallet has.
+        #
+        # @param [String] wallet_id - The wallet id that is offered by `create_wallet()` method.
+        # @param [Boolean] only_finalized - The balance includes only finalized UTXO value if it
+        #                                   is true. Default is false.
+        # @return [Integer] The balance of the wallet. The unit is 'tapyrus'.
+        #                   1 TPC equals to 10^8 'tapyrus'.
+        def balance(wallet_id, only_finalized = false)
+          raise NotImplementedError, "You must implement #{self.class}##{__method__}"
+        end
+
+        # Returns all the UTXOs that the wallet has.
+        #
+        # @param [String] wallet_id - The wallet id that is offered by `create_wallet()` method.
+        # @param [Boolean] only_finalized - The UTXOs includes only finalized UTXO value if it
+        #                                   is true. Default is false.
+        # @return [Array of UTXO]
+        #
+        # ## The UTXO structure
+        #
+        # - txid: [String] Transaction id
+        # - vout: [Integer] Output index
+        # - amount: [Integer] Amount of the UTXO as tapyrus unit
+        # - finalized: [Boolean] Whether the UTXO is finalized
+        def list_unspent(wallet_id, only_finalized = true)
+          raise NotImplementedError, "You must implement #{self.class}##{__method__}"
+        end
+
+        # Sign to the transaction with a key in the wallet.
+        #
+        # @param [String] wallet_id - The wallet id that is offered by `create_wallet()` method.
+        # @param [Tapyrus::Tx] tx - The Transaction will be signed.
+        # @return [Tapyrus::Tx]
+        def sign_tx(wallet_id, tx)
+          raise NotImplementedError, "You must implement #{self.class}##{__method__}"
+        end
+
+        # Returns an address to receive coin.
+        #
+        # @param [String] wallet_id - The wallet id that is offered by `create_wallet()` method.
+        # @return [String] P2PKH address
+        def receive_address(wallet_id)
+          raise NotImplementedError, "You must implement #{self.class}##{__method__}"
+        end
+
+        # Returns an address to change coin.
+        #
+        # @param [String] wallet_id - The wallet id that is offered by `create_wallet()` method.
+        # @return [String] P2PKH address
+        def change_address(wallet_id)
+          raise NotImplementedError, "You must implement #{self.class}##{__method__}"
+        end
+
+        # Returns a public key.
+        #
+        # @param [String] wallet_id - The wallet id that is offered by `create_wallet()` method.
+        # @return [Tapyrus::Key]
+        def pubkey(wallet_id)
+          raise NotImplementedError, "You must implement #{self.class}##{__method__}"
+        end
+      end
+    end
+  end
+end
