+++ lib/glueby/internal/wallet/errors.rb
@@ -1,0 +1,10 @@
+module Glueby
+  module Internal
+    class Wallet
+      module Errors
+        class ShouldInitializeWalletAdapter < StandardError; end
+        class WalletUnloaded < StandardError; end
+      end
+    end
+  end
+end
