--- lib/glueby/wallet/errors.rb
@@ -1,8 +1,0 @@
-module Glueby
-  class Wallet
-    module Errors
-      class ShouldInitializeWalletAdapter < StandardError; end
-      class WalletUnloaded < StandardError; end
-    end
-  end
-end
